requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
earthengine-api>=1.4.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
shapely>=2.0.0
requests>=2.31.0
pyproj>=3.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.9
